ARM: OMAP2+: AM33xx: Enhance McASP FIFO length
Existing fifo level of 1 being configured for McASP is as good as
fifo being disabled. Increase the fifo level so as to fully utilize
McASP FIFO support.

The maximum fifo level possible is 64. It this patch, it is being
set to the midway point 32 so that there is some fifo available
for new data when the level is reached.

Signed-off-by: Hebbar Gururaja <gururaja.hebbar@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>