usb: musb: ti81xx: Check for context loss during resume
Context restore is not required when there is no loss of context which
is the case with standby.

This patch adds support to check for loss of context and context restore
is done if there has been a loss of context.

This reduces the overall resume latency of Standby approximately by
200ms.

Signed-off-by: Satyanarayana Sandhya <sandhya.satyanarayana@ti.com>
Signed-off-by: ShankarMurthy, Akshay <akshay.s@ti.com>