ARM: OMAP2+: AM335x: Handle OPP50 Bootup
As per Advisory 1.0.15 (ARM Cortex-A8: OPP50 Operation on MPU Domain Not
Supported), when the board is booted with OPP50, reliable operation is
not guaranteed for OPP greater than OPP100 (OPP120, TURBO, NITRO).

So, Check if the board is booted at OPP50 voltage & if yes, disable
higher OPP (OPP120, TURBO, NITRO).

Signed-off-by: Hebbar Gururaja <gururaja.hebbar@ti.com>