Revert "ARM: OMAP: AM33XX: Add OPP table for PG2.1 AM335x"
This reverts commit ee9dfd8d729d3e7b5ce9e404a0e87f27f6f79135.

This patch checks for the package type for checking the supported opp
bits & also if the bits are set, the opp table is updated.
However, checking package type bit is not required & also, the opp bit
checking must be reversed.

A fix for the same will follow after this commit